backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
